# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehaus API",
        "description": "REST API v1 untuk sistem manajemen inventaris Warehaus. Setiap endpoint mengembalikan envelope { success, message, data }.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Warehaus API server"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List categories with product counts",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a single category",
                "operationId": "d5023633bd2ab475b5f7d3fe21113fd9",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Aggregate KPIs and 12-month borrowing trend",
                "operationId": "00464eac0b2daa2b7dfd173ce52781f0",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiEnvelope"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "total_stock": {
                                                            "type": "integer"
                                                        },
                                                        "borrowed_count": {
                                                            "type": "integer"
                                                        },
                                                        "available_stock": {
                                                            "type": "integer"
                                                        },
                                                        "total_categories": {
                                                            "type": "integer"
                                                        },
                                                        "total_products": {
                                                            "type": "integer"
                                                        },
                                                        "low_stock_products": {
                                                            "type": "integer"
                                                        },
                                                        "active_borrowings": {
                                                            "type": "integer"
                                                        },
                                                        "overdue_borrowings": {
                                                            "type": "integer"
                                                        },
                                                        "monthly_trends": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "month": {
                                                                        "type": "string",
                                                                        "example": "2026-07"
                                                                    },
                                                                    "total": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "description": "Base OpenAPI info + security scheme for the Warehaus API.\n\nKept as an empty controller purely so the l5-swagger annotation scanner\nhas a well-known place to pick up the top-level metadata.",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validasi gagal."
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login via POST /api/v1/login to obtain a token.",
                "bearerFormat": "Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}